+++ src/demo/platform/file.c
@@ -223,3 +223,41 @@ tb_int_t tb_demo_platform_file_main(tb_int_t argc, tb_char_t** argv)
     return 0;
 }
 
+/* //////////////////////////////////////////////////////////////////////////////////////
+ * implementation
+ */
+tb_int_t tb_demo_platform_file_signature_main(tb_int_t argc, tb_char_t** argv)
+{
+    tb_file_signature_info_t info = {0};
+    tb_char_t const* path = argv[1];
+
+    // default to notepad.exe on windows if no path
+    // default to argv[0] on other platforms
+    if (!path) 
+    {
+#ifdef TB_CONFIG_OS_WINDOWS
+        path = "C:\\Windows\\System32\\notepad.exe";
+#else
+        path = argv[0];
+#endif
+    }
+
+    // check signature
+    if (tb_file_get_signature_info(path, &info))
+    {
+        tb_trace_i("path: %s", path);
+        tb_trace_i("is_signed: %s", info.is_signed? "yes" : "no");
+        tb_trace_i("is_trusted: %s", info.is_trusted? "yes" : "no");
+        if (info.is_signed)
+        {
+            tb_trace_i("signer_name: %s", info.signer_name);
+        }
+    }
+    else
+    {
+        tb_trace_e("failed to get signature info: %s", path);
+    }
+    return 0;
+}
+
+
