--- src/tbox/platform/file_signature.c
@@ -1,46 +1,0 @@
-/*!The Treasure Box Library
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Copyright (C) 2009-present, TBOOX Open Source Group.
- *
- * @author      ruki
- * @file        file_signature.c
- * @ingroup     platform
- *
- */
-
-/* //////////////////////////////////////////////////////////////////////////////////////
- * trace
- */
-#define TB_TRACE_MODULE_NAME                "platform_file_signature"
-#define TB_TRACE_MODULE_DEBUG               (0)
-
-/* //////////////////////////////////////////////////////////////////////////////////////
- * includes
- */
-#include "prefix.h"
-#include "file.h"
-
-/* //////////////////////////////////////////////////////////////////////////////////////
- * implementation
- */
-#if defined(TB_CONFIG_OS_WINDOWS)
-#   include "windows/file_signature.c"
-#else
-tb_bool_t tb_file_get_signature_info(tb_char_t const* path, tb_file_signature_info_t* info)
-{
-    tb_trace_noimpl();
-    return tb_false;
-}
-#endif
